Fix DRY violations in pipeline dashboard components
- Move TasksChangedPayload type to pipeline-dashboard-utils (was duplicated in both dashboard files)
- Extract formatMs to utils and refactor formatElapsed to reuse it (near-identical logic)
- Remove dead code branch: colCounts.size === 0 check inside activeTasks.length > 0 guard
